[feat] layout 파일 생성
